2022-10-25 Lambda 文件大小 概double

+++ erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/DeliveryDocsDTO.java
@@ -92,7 +92,7 @@ public class DeliveryDocsDTO implements Serializable {
     /**
      * 文件大小
      */
-    private Integer fileSize;
+    private Double fileSize;
 
 
     /**
